removing event coordinator
