+++ packages/dev/core/src/FlowGraph/flowGraphContext.ts
@@ -4,9 +4,9 @@ import type { Scene } from "../scene";
 import type { FlowGraphAsyncExecutionBlock } from "./flowGraphAsyncExecutionBlock";
 import type { FlowGraphBlock } from "./flowGraphBlock";
 import type { FlowGraphDataConnection } from "./flowGraphDataConnection";
-import type { FlowGraphEventCoordinator } from "./flowGraphEventCoordinator";
 import type { FlowGraph } from "./flowGraph";
 import { defaultValueParseFunction, defaultValueSerializationFunction } from "./serialization";
+import type { FlowGraphCoordinator } from "./flowGraphCoordinator";
 
 /**
  * Construction parameters for the context.
@@ -20,7 +20,7 @@ export interface IFlowGraphContextConfiguration {
     /**
      * The event coordinator used by the flow graph context.
      */
-    readonly eventCoordinator: FlowGraphEventCoordinator;
+    readonly coordinator: FlowGraphCoordinator;
 }
 /**
  * @experimental
