--- packages/dev/core/src/FlowGraph/flowGraphCustomEvent.ts
@@ -1,5 +1,0 @@
-/**
- * @experimental
- * Represents a custom event
- */
-export type FlowGraphCustomEvent = any;
